refactor: 일급 컬렉션 vo로 이동
