+++ src/test/java/com/onebyte/life4cut/picture/domain/PictureTagRelationsTest.java
@@ -3,6 +3,7 @@ package com.onebyte.life4cut.picture.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.onebyte.life4cut.fixture.PictureTagRelationFixtureFactory;
+import com.onebyte.life4cut.picture.domain.vo.PictureTagRelations;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
