+++ src/test/java/com/onebyte/life4cut/picture/service/PictureServiceTest.java
@@ -22,8 +22,8 @@ import com.onebyte.life4cut.fixture.UserAlbumFixtureFactory;
 import com.onebyte.life4cut.picture.domain.Picture;
 import com.onebyte.life4cut.picture.repository.PictureRepository;
 import com.onebyte.life4cut.pictureTag.domain.PictureTag;
-import com.onebyte.life4cut.pictureTag.domain.PictureTags;
 import com.onebyte.life4cut.pictureTag.domain.vo.PictureTagName;
+import com.onebyte.life4cut.pictureTag.domain.vo.PictureTags;
 import com.onebyte.life4cut.pictureTag.repository.PictureTagRepository;
 import com.onebyte.life4cut.slot.domain.Slot;
 import com.onebyte.life4cut.slot.exception.SlotNotFoundException;
